+++ bfa_c/test/cram/multi-files/test2.t/file2.c
@@ -3,6 +3,6 @@
 int main()
 {
   int ret = compute();
-  __assert__(ret == 10);
+  __assert__(ret == 12);
   return 0;
 }
