+++ bfa_c/test/cram/multi-files/test2.t/test2.h
@@ -1,0 +1,3 @@
+int __assert__(int x);
+
+int compute();
