+++ bfa_c/test/cram/multi-files/test3.t/file1.c
@@ -1,0 +1,6 @@
+#include "test3.h"
+
+int add(int x, int y)
+{
+    return x + y;
+}
