+++ bfa_c/test/cram/multi-files/test3.t/test3.h
@@ -1,0 +1,8 @@
+#ifndef TEST3_H
+#define TEST3_H
+
+int __assert__(int x);
+
+int add(int x, int y);
+
+#endif // TEST3_H
