--- bfa_c/test/cram/multi-files/test3/file1.c
@@ -1,7 +1,0 @@
-#include "test3.h"
-
-// ...existing code...
-int operate(operation op, int a, int b)
-{
-    return op(a, b);
-}
