--- bfa_c/test/cram/multi-files/test3/run.t
@@ -1,3 +1,0 @@
-  $ bfa-c show-ail file1.c file2.c -I .
-
-  $ bfa-c exec-main file1.c file2.c -I .
