--- bfa_c/test/cram/multi-files/test3/test3.h
@@ -1,7 +1,0 @@
-#ifndef TEST3_H
-#define TEST3_H
-
-typedef int (*operation)(int, int);
-int operate(operation op, int a, int b);
-
-#endif // TEST3_H
