+++ bfa_c/test/cram/multi-files/test4.t/file1.c
@@ -1,3 +1,5 @@
+#include "test4.h"
+
 int funcB(int x)
 {
   if (x <= 0)
@@ -9,6 +11,6 @@ int funcB(int x)
 
 int main()
 {
-  assert(funcA(5) == 1);
+  __assert__(funcA(5) == 1);
   return 0;
 }
