+++ bfa_c/test/cram/multi-files/test4.t/file2.c
@@ -1,3 +1,5 @@
+#include "test4.h"
+
 int funcA(int x)
 {
   if (x <= 0)
