+++ bfa_c/test/cram/multi-files/test4.t/test4.h
@@ -1,7 +1,8 @@
 #ifndef TEST4_H
 #define TEST4_H
 
-int funcA(int); // changed return type to int
-int funcB(int); // changed return type to int
+int __assert__(int);
+int funcA(int);
+int funcB(int);
 
 #endif // TEST4_H
