+++ bfa_c/test/cram/multi-files/test5.t/file3.c
@@ -1,3 +1,5 @@
+#include "test5.h"
+
 int main()
 {
   int ret = fnA() + fnB();
