+++ bfa_c/test/cram/multi-files/test5.t/test5.h
@@ -1,0 +1,8 @@
+#ifndef TEST5_H
+#define TEST5_H
+
+int __assert__(int);
+int fnA();
+int fnB();
+
+#endif // TEST5_H
